+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added extra filtering methods for ElementList
+
 ## [0.12.0] - 2023-11-10
 
 ### Added